requirements.txt
python-dotenv~=1.0.0
langchain~=0.0.350
autopep8~=2.0.4
openai~=1.5.0
numpy~=1.24.2
torch~=2.0.0
transformers~=4.36.1